Fixed broken SET BUTTON command in operator module.
git-svn-id: http://www.bci2000.org/svn/trunk@4606 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/core/Operator/OperatorQt/MainWindow.h
@@ -93,7 +93,7 @@ class MainWindow : public QMainWindow
   void GetParameters();
 
  public slots:
-  void SetFunctionButton( size_t idx, const QString& title, const QString& script );
+  void SetFunctionButton( int idx, const QString& title, const QString& script );
   void PutParameters();
 
  private:
